// src/api/generated/models/Task.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Manager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TaskPriority } from './TaskPriority';
import {
    TaskPriorityFromJSON,
    TaskPriorityFromJSONTyped,
    TaskPriorityToJSON,
    TaskPriorityToJSONTyped,
} from './TaskPriority';
import type { TaskStatus } from './TaskStatus';
import {
    TaskStatusFromJSON,
    TaskStatusFromJSONTyped,
    TaskStatusToJSON,
    TaskStatusToJSONTyped,
} from './TaskStatus';

/**
 * 
 * @export
 * @interface Task
 */
export interface Task {
    /**
     * 
     * @type {string}
     * @memberof Task
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Task
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Task
     */
    description?: string | null;
    /**
     * 
     * @type {TaskStatus}
     * @memberof Task
     */
    status: TaskStatus;
    /**
     * 
     * @type {TaskPriority}
     * @memberof Task
     */
    priority: TaskPriority;
    /**
     * 
     * @type {Date}
     * @memberof Task
     */
    deadline?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof Task
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Task
     */
    updatedAt: Date;
}



/**
 * Check if a given object implements the Task interface.
 */
export function instanceOfTask(value: object): value is Task {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('priority' in value) || value['priority'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function TaskFromJSON(json: any): Task {
    return TaskFromJSONTyped(json, false);
}

export function TaskFromJSONTyped(json: any, ignoreDiscriminator: boolean): Task {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'status': TaskStatusFromJSON(json['status']),
        'priority': TaskPriorityFromJSON(json['priority']),
        'deadline': json['deadline'] == null ? undefined : (new Date(json['deadline'])),
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function TaskToJSON(json: any): Task {
    return TaskToJSONTyped(json, false);
}

export function TaskToJSONTyped(value?: Task | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'description': value['description'],
        'status': TaskStatusToJSON(value['status']),
        'priority': TaskPriorityToJSON(value['priority']),
        'deadline': value['deadline'] == null ? value['deadline'] : value['deadline'].toISOString(),
        'createdAt': value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'].toISOString(),
    };
}